+++ src/field_ext.rs
@@ -1,4 +1,4 @@
-use syn::{Field, Ident, NestedMeta, Path, PathSegment, Type, TypePath};
+use syn::{Field, Ident, Meta, Path, PathSegment, Type, TypePath};
 
 use crate::util;
 
@@ -36,7 +36,7 @@ pub trait FieldExt {
     /// # Panics
     ///
     /// Panics if there is more than one parameter for the tag.
-    fn tag_parameter(&self, namespace: &Path, tag: &Path) -> Option<NestedMeta>;
+    fn tag_parameter(&self, namespace: &Path, tag: &Path) -> Option<Meta>;
 
     /// Returns the parameters from `#[namespace(tag(param1, param2, ..))]`.
     ///
@@ -44,7 +44,7 @@ pub trait FieldExt {
     ///
     /// * `namespace`: The `path()` of the first-level attribute.
     /// * `tag`: The `path()` of the second-level attribute.
-    fn tag_parameters(&self, namespace: &Path, tag: &Path) -> Vec<NestedMeta>;
+    fn tag_parameters(&self, namespace: &Path, tag: &Path) -> Vec<Meta>;
 }
 
 impl FieldExt for Field {
@@ -73,18 +73,18 @@ impl FieldExt for Field {
         util::contains_tag(&self.attrs, namespace, tag)
     }
 
-    fn tag_parameter(&self, namespace: &Path, tag: &Path) -> Option<NestedMeta> {
+    fn tag_parameter(&self, namespace: &Path, tag: &Path) -> Option<Meta> {
         util::tag_parameter(&self.attrs, namespace, tag)
     }
 
-    fn tag_parameters(&self, namespace: &Path, tag: &Path) -> Vec<NestedMeta> {
+    fn tag_parameters(&self, namespace: &Path, tag: &Path) -> Vec<Meta> {
         util::tag_parameters(&self.attrs, namespace, tag)
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use syn::{parse_quote, Fields, FieldsNamed, Lit, Meta, NestedMeta};
+    use syn::{parse_quote, Fields, FieldsNamed, Meta, MetaNameValue};
 
     use super::FieldExt;
 
@@ -145,14 +145,12 @@ mod tests {
 
         assert_eq!(
             field.tag_parameter(&parse_quote!(my::derive), &parse_quote!(tag::name)),
-            Some(NestedMeta::Meta(Meta::Path(parse_quote!(Magic))))
+            Some(Meta::Path(parse_quote!(Magic)))
         );
     }
 
     #[test]
-    #[should_panic(
-        expected = "Expected exactly one identifier for `#[my::derive(tag::name(..))]`."
-    )]
+    #[should_panic(expected = "Expected exactly one parameter for `#[my::derive(tag::name(..))]`.")]
     fn tag_parameter_panics_when_multiple_parameters_present() {
         let fields_named: FieldsNamed = parse_quote! {{
             #[my::derive(tag::name(Magic::One, Magic::Two))]
@@ -175,14 +173,14 @@ mod tests {
 
         assert_eq!(
             field.tag_parameters(&parse_quote!(my::derive), &parse_quote!(tag::name)),
-            Vec::<NestedMeta>::new()
+            Vec::<Meta>::new()
         );
     }
 
     #[test]
     fn tag_parameters_returns_paths_when_present() {
         let fields_named: FieldsNamed = parse_quote! {{
-            #[my::derive(tag::name(Magic::One, "{ Magic::Two }"))]
+            #[my::derive(tag::name(Magic::One, second = "{ Magic::Two }"))]
             pub name: u32,
         }};
         let fields = Fields::from(fields_named);
@@ -191,8 +189,12 @@ mod tests {
         assert_eq!(
             field.tag_parameters(&parse_quote!(my::derive), &parse_quote!(tag::name)),
             vec![
-                NestedMeta::Meta(Meta::Path(parse_quote!(Magic::One))),
-                NestedMeta::Lit(Lit::Str(parse_quote!("{ Magic::Two }"))),
+                Meta::Path(parse_quote!(Magic::One)),
+                Meta::NameValue(MetaNameValue {
+                    path: parse_quote!(second),
+                    eq_token: Default::default(),
+                    value: parse_quote!("{ Magic::Two }")
+                }),
             ]
         );
     }